extract RecipientInput into a separate component
